nearly done with the new mustache templating - still need to add sOIL and sAAVE and update the readme
