+++ mustache/templates/exchanger/set-start-blocks.js
@@ -1,0 +1,42 @@
+module.exports = {
+  Exchanger: {
+    prod: 10557958,
+    test: null,
+  },
+  Exchanger_v2: {
+    prod: 10772592,
+    test: null,
+  },
+  Exchanger_v3: {
+    prod: 11012438,
+    test: null,
+  },
+  Synthetix: {
+    prod: 10782000,
+    test: null,
+  },
+  RatesBlockOne: {
+    prod: 10537958,
+    test: null,
+  },
+  RatesBlockTwo: {
+    prod: 10773070,
+    test: null,
+  },
+  RatesBlockThree: {
+    prod: 10873070,
+    test: null,
+  },
+  RatesBlockFour: {
+    prod: 10921231,
+    test: null,
+  },
+  RatesBlockFive: {
+    prod: 10950006,
+    test: null,
+  },
+  RatesBlockSix: {
+    prod: 10960006,
+    test: null,
+  },
+};
