--- mustache/templates/rates/modify_start_blocks_for_prod.js
@@ -1,72 +1,0 @@
-// Use this file to update the prod start blocks for the rates subgraph
-module.exports = {
-  ExchangeRates_v231: 7626469,
-  ExchangeRates_v240: 8075694,
-  ExchangeRates_v272: 8120141,
-  ExchangeRates_v210: 8622895,
-  ExchangeRates_v213: 8971442,
-  ExchangeRates_v217: 9123410,
-  ExchangeRates_v219: 9518289,
-  ExchangeRates_v223: 10364342,
-  ExchangeRates: 10773070,
-  AggregatorAUD: 9085450,
-  AggregatorEUR: 9085417,
-  AggregatorCHF: 9085502,
-  AggregatorGBP: 9085494,
-  AggregatorJPY: 9085433,
-  AggregatorXAG: 9085563,
-  AggregatorXAU: 9085543,
-  AggregatorFTSE: 9755177,
-  AggregatorN225: 9755246,
-  AggregatorSNX: 10773070,
-  AggregatorETH: 10773070,
-  AggregatorCOMP: 10773070,
-  AggregatorKNC: 10773070,
-  AggregatorLEND: 10773070,
-  AggregatorREN: 10773070,
-  AggregatorBTC: 10773070,
-  AggregatorBNB: 10773070,
-  AggregatorTRX: 10773070,
-  AggregatorXTZ: 10773070,
-  AggregatorXRP: 10773070,
-  AggregatorLTC: 10773070,
-  AggregatorLINK: 10773070,
-  AggregatorEOS: 10773070,
-  AggregatorBCH: 10773070,
-  AggregatorETC: 10773070,
-  AggregatorDASH: 10773070,
-  AggregatorXMR: 10773070,
-  AggregatorADA: 10773070,
-  AggregatorCEX: 10773070,
-  AggregatorDEFI: 10773070,
-  AggregatorXAU_3: 10873070,
-  AggregatorXAG_3: 10873070,
-  AggregatorLINK_3: 10921231,
-  AggregatorBTC_3: 10921231,
-  AggregatorAUD_3: 10921231,
-  AggregatorEUR_3: 10921231,
-  AggregatorBCH_3: 10950006,
-  AggregatorEOS_3: 10950006,
-  AggregatorBNB_3: 10950006,
-  AggregatorADA_3: 10950006,
-  AggregatorREN_3: 10950006,
-  AggregatorLEND_3: 10950006,
-  AggregatorKNC_3: 10950006,
-  AggregatorSNX_3: 10950006,
-  AggregatorDASH_3: 10960006,
-  AggregatorETC_3: 10960006,
-  AggregatorLTC_3: 10960006,
-  AggregatorXMR_3: 10960006,
-  AggregatorXRP_3: 10960006,
-  AggregatorTRX_3: 10960006,
-  AggregatorXTZ_3: 10960006,
-  AggregatorCOMP_3: 10960006,
-  AggregatorDEFI_3: 10960006,
-  AggregatorCEX_3: 10960006,
-  AggregatorETH_3: 10960006,
-  AggregatorCHF_3: 10960006,
-  AggregatorJPY_3: 10960006,
-  AggregatorGBP_3: 10960006,
-  AggregatorFTSE_3: 10960006,
-  AggregatorNIKKEI_3: 10960006,
-};
