--- mustache/templates/rates/modify_start_blocks_for_testing.js
@@ -1,79 +1,0 @@
-/**
- * Use this file to update the prod start blocks for the rates subgraph
- * you can also set them all to a single block using the
- * UNIVERSAL_START_BLOCK environment variable in package json
- * NOTE: if you change the test blocks you have to run `npm run codegen:rates:test` command
- * or else it will use the prod settings
- * */
-
-module.exports = {
-  ExchangeRates_v231: null,
-  ExchangeRates_v240: null,
-  ExchangeRates_v272: null,
-  ExchangeRates_v210: null,
-  ExchangeRates_v213: null,
-  ExchangeRates_v217: null,
-  ExchangeRates_v219: null,
-  ExchangeRates_v223: null,
-  ExchangeRates: null,
-  AggregatorAUD: null,
-  AggregatorEUR: null,
-  AggregatorCHF: null,
-  AggregatorGBP: null,
-  AggregatorJPY: null,
-  AggregatorXAG: null,
-  AggregatorXAU: null,
-  AggregatorFTSE: null,
-  AggregatorN225: null,
-  AggregatorSNX: null,
-  AggregatorETH: null,
-  AggregatorCOMP: null,
-  AggregatorKNC: null,
-  AggregatorLEND: null,
-  AggregatorREN: null,
-  AggregatorBTC: null,
-  AggregatorBNB: null,
-  AggregatorTRX: null,
-  AggregatorXTZ: null,
-  AggregatorXRP: null,
-  AggregatorLTC: null,
-  AggregatorLINK: null,
-  AggregatorEOS: null,
-  AggregatorBCH: null,
-  AggregatorETC: null,
-  AggregatorDASH: null,
-  AggregatorXMR: null,
-  AggregatorADA: null,
-  AggregatorCEX: null,
-  AggregatorDEFI: null,
-  AggregatorXAU_3: null,
-  AggregatorXAG_3: null,
-  AggregatorLINK_3: null,
-  AggregatorBTC_3: null,
-  AggregatorAUD_3: null,
-  AggregatorEUR_3: null,
-  AggregatorBCH_3: null,
-  AggregatorEOS_3: null,
-  AggregatorBNB_3: null,
-  AggregatorADA_3: null,
-  AggregatorREN_3: null,
-  AggregatorLEND_3: null,
-  AggregatorKNC_3: null,
-  AggregatorSNX_3: null,
-  AggregatorDASH_3: null,
-  AggregatorETC_3: null,
-  AggregatorLTC_3: null,
-  AggregatorXMR_3: null,
-  AggregatorXRP_3: null,
-  AggregatorTRX_3: null,
-  AggregatorXTZ_3: null,
-  AggregatorCOMP_3: null,
-  AggregatorDEFI_3: null,
-  AggregatorCEX_3: null,
-  AggregatorETH_3: null,
-  AggregatorCHF_3: null,
-  AggregatorJPY_3: null,
-  AggregatorGBP_3: null,
-  AggregatorFTSE_3: null,
-  AggregatorNIKKEI_3: null,
-};
