+++ mustache/templates/rates/set-start-blocks.js
@@ -1,0 +1,349 @@
+const Exchanger = require('../exchanger/set-start-blocks');
+
+module.exports = {
+  ExchangeRates_v231: {
+    prod: 7626469,
+    test: null,
+    exchanger: Exchanger.RatesBlockOne,
+  },
+  ExchangeRates_v240: {
+    prod: 8075694,
+    test: null,
+    exchanger: Exchanger.RatesBlockOne,
+  },
+  ExchangeRates_v272: {
+    prod: 8120141,
+    test: null,
+    exchanger: Exchanger.RatesBlockOne,
+  },
+  ExchangeRates_v210: {
+    prod: 8622895,
+    test: null,
+    exchanger: Exchanger.RatesBlockOne,
+  },
+  ExchangeRates_v213: {
+    prod: 8971442,
+    test: null,
+    exchanger: Exchanger.RatesBlockOne,
+  },
+  ExchangeRates_v217: {
+    prod: 9123410,
+    test: null,
+    exchanger: Exchanger.RatesBlockOne,
+  },
+  ExchangeRates_v219: {
+    prod: 9518289,
+    test: null,
+    exchanger: Exchanger.RatesBlockOne,
+  },
+  ExchangeRates_v223: {
+    prod: 10364342,
+    test: null,
+    exchanger: Exchanger.RatesBlockOne,
+  },
+  ExchangeRates: {
+    prod: 10773070,
+    test: null,
+    exchanger: Exchanger.RatesBlockTwo,
+  },
+  AggregatorAUD: {
+    prod: 9085450,
+    test: null,
+    exchanger: Exchanger.RatesBlockOne,
+  },
+  AggregatorEUR: {
+    prod: 9085417,
+    test: null,
+    exchanger: Exchanger.RatesBlockOne,
+  },
+  AggregatorCHF: {
+    prod: 9085502,
+    test: null,
+    exchanger: Exchanger.RatesBlockOne,
+  },
+  AggregatorGBP: {
+    prod: 9085494,
+    test: null,
+    exchanger: Exchanger.RatesBlockOne,
+  },
+  AggregatorJPY: {
+    prod: 9085433,
+    test: null,
+    exchanger: Exchanger.RatesBlockOne,
+  },
+  AggregatorXAG: {
+    prod: 9085563,
+    test: null,
+    exchanger: Exchanger.RatesBlockOne,
+  },
+  AggregatorXAU: {
+    prod: 9085543,
+    test: null,
+    exchanger: Exchanger.RatesBlockOne,
+  },
+  AggregatorFTSE: {
+    prod: 9755177,
+    test: null,
+    exchanger: Exchanger.RatesBlockOne,
+  },
+  AggregatorN225: {
+    prod: 9755246,
+    test: null,
+    exchanger: Exchanger.RatesBlockOne,
+  },
+  AggregatorSNX: {
+    prod: 10773070,
+    test: null,
+    exchanger: Exchanger.RatesBlockTwo,
+  },
+  AggregatorETH: {
+    prod: 10773070,
+    test: null,
+    exchanger: Exchanger.RatesBlockTwo,
+  },
+  AggregatorCOMP: {
+    prod: 10773070,
+    test: null,
+    exchanger: Exchanger.RatesBlockTwo,
+  },
+  AggregatorKNC: {
+    prod: 10773070,
+    test: null,
+    exchanger: Exchanger.RatesBlockTwo,
+  },
+  AggregatorLEND: {
+    prod: 10773070,
+    test: null,
+    exchanger: Exchanger.RatesBlockTwo,
+  },
+  AggregatorREN: {
+    prod: 10773070,
+    test: null,
+    exchanger: Exchanger.RatesBlockTwo,
+  },
+  AggregatorBTC: {
+    prod: 10773070,
+    test: null,
+    exchanger: Exchanger.RatesBlockTwo,
+  },
+  AggregatorBNB: {
+    prod: 10773070,
+    test: null,
+    exchanger: Exchanger.RatesBlockTwo,
+  },
+  AggregatorTRX: {
+    prod: 10773070,
+    test: null,
+    exchanger: Exchanger.RatesBlockTwo,
+  },
+  AggregatorXTZ: {
+    prod: 10773070,
+    test: null,
+    exchanger: Exchanger.RatesBlockTwo,
+  },
+  AggregatorXRP: {
+    prod: 10773070,
+    test: null,
+    exchanger: Exchanger.RatesBlockTwo,
+  },
+  AggregatorLTC: {
+    prod: 10773070,
+    test: null,
+    exchanger: Exchanger.RatesBlockTwo,
+  },
+  AggregatorLINK: {
+    prod: 10773070,
+    test: null,
+    exchanger: Exchanger.RatesBlockTwo,
+  },
+  AggregatorEOS: {
+    prod: 10773070,
+    test: null,
+    exchanger: Exchanger.RatesBlockTwo,
+  },
+  AggregatorBCH: {
+    prod: 10773070,
+    test: null,
+    exchanger: Exchanger.RatesBlockTwo,
+  },
+  AggregatorETC: {
+    prod: 10773070,
+    test: null,
+    exchanger: Exchanger.RatesBlockTwo,
+  },
+  AggregatorDASH: {
+    prod: 10773070,
+    test: null,
+    exchanger: Exchanger.RatesBlockTwo,
+  },
+  AggregatorXMR: {
+    prod: 10773070,
+    test: null,
+    exchanger: Exchanger.RatesBlockTwo,
+  },
+  AggregatorADA: {
+    prod: 10773070,
+    test: null,
+    exchanger: Exchanger.RatesBlockTwo,
+  },
+  AggregatorCEX: {
+    prod: 10773070,
+    test: null,
+    exchanger: Exchanger.RatesBlockTwo,
+  },
+  AggregatorDEFI: {
+    prod: 10773070,
+    test: null,
+    exchanger: Exchanger.RatesBlockTwo,
+  },
+  AggregatorXAU_3: {
+    prod: 10873070,
+    test: null,
+    exchanger: Exchanger.RatesBlockThree,
+  },
+  AggregatorXAG_3: {
+    prod: 10873070,
+    test: null,
+    exchanger: Exchanger.RatesBlockThree,
+  },
+  AggregatorLINK_3: {
+    prod: 10921231,
+    test: null,
+    exchanger: Exchanger.RatesBlockFour,
+  },
+  AggregatorBTC_3: {
+    prod: 10921231,
+    test: null,
+    exchanger: Exchanger.RatesBlockFour,
+  },
+  AggregatorAUD_3: {
+    prod: 10921231,
+    test: null,
+    exchanger: Exchanger.RatesBlockFour,
+  },
+  AggregatorEUR_3: {
+    prod: 10921231,
+    test: null,
+    exchanger: Exchanger.RatesBlockFour,
+  },
+  AggregatorBCH_3: {
+    prod: 10950006,
+    test: null,
+    exchanger: Exchanger.RatesBlockFive,
+  },
+  AggregatorEOS_3: {
+    prod: 10950006,
+    test: null,
+    exchanger: Exchanger.RatesBlockFive,
+  },
+  AggregatorBNB_3: {
+    prod: 10950006,
+    test: null,
+    exchanger: Exchanger.RatesBlockFive,
+  },
+  AggregatorADA_3: {
+    prod: 10950006,
+    test: null,
+    exchanger: Exchanger.RatesBlockFive,
+  },
+  AggregatorREN_3: {
+    prod: 10950006,
+    test: null,
+    exchanger: Exchanger.RatesBlockFive,
+  },
+  AggregatorLEND_3: {
+    prod: 10950006,
+    test: null,
+    exchanger: Exchanger.RatesBlockFive,
+  },
+  AggregatorKNC_3: {
+    prod: 10950006,
+    test: null,
+    exchanger: Exchanger.RatesBlockFive,
+  },
+  AggregatorSNX_3: {
+    prod: 10950006,
+    test: null,
+    exchanger: Exchanger.RatesBlockFive,
+  },
+  AggregatorDASH_3: {
+    prod: 10960006,
+    test: null,
+    exchanger: Exchanger.RatesBlockSix,
+  },
+  AggregatorETC_3: {
+    prod: 10960006,
+    test: null,
+    exchanger: Exchanger.RatesBlockSix,
+  },
+  AggregatorLTC_3: {
+    prod: 10960006,
+    test: null,
+    exchanger: Exchanger.RatesBlockSix,
+  },
+  AggregatorXMR_3: {
+    prod: 10960006,
+    test: null,
+    exchanger: Exchanger.RatesBlockSix,
+  },
+  AggregatorXRP_3: {
+    prod: 10960006,
+    test: null,
+    exchanger: Exchanger.RatesBlockSix,
+  },
+  AggregatorTRX_3: {
+    prod: 10960006,
+    test: null,
+    exchanger: Exchanger.RatesBlockSix,
+  },
+  AggregatorXTZ_3: {
+    prod: 10960006,
+    test: null,
+    exchanger: Exchanger.RatesBlockSix,
+  },
+  AggregatorCOMP_3: {
+    prod: 10960006,
+    test: null,
+    exchanger: Exchanger.RatesBlockSix,
+  },
+  AggregatorDEFI_3: {
+    prod: 10960006,
+    test: null,
+    exchanger: Exchanger.RatesBlockSix,
+  },
+  AggregatorCEX_3: {
+    prod: 10960006,
+    test: null,
+    exchanger: Exchanger.RatesBlockSix,
+  },
+  AggregatorETH_3: {
+    prod: 10960006,
+    test: null,
+    exchanger: Exchanger.RatesBlockSix,
+  },
+  AggregatorCHF_3: {
+    prod: 10960006,
+    test: null,
+    exchanger: Exchanger.RatesBlockSix,
+  },
+  AggregatorJPY_3: {
+    prod: 10960006,
+    test: null,
+    exchanger: Exchanger.RatesBlockSix,
+  },
+  AggregatorGBP_3: {
+    prod: 10960006,
+    test: null,
+    exchanger: Exchanger.RatesBlockSix,
+  },
+  AggregatorFTSE_3: {
+    prod: 10960006,
+    test: null,
+    exchanger: Exchanger.RatesBlockSix,
+  },
+  AggregatorNIKKEI_3: {
+    prod: 10960006,
+    test: null,
+    exchanger: Exchanger.RatesBlockSix,
+  },
+};
