+++ package.json
@@ -10,19 +10,17 @@
   },
   "scripts": {
     "lint": "eslint .",
-    "yaml:rates": "node create-yaml-json rates && mustache mustache/yaml_output/synthetix-rates.json mustache/yaml.mustache > subgraphs/synthetix-rates.yaml",
-    "yaml:exchanger": "node create-yaml-json exchanger && mustache mustache/yaml_output/synthetix-exchanger.json mustache/yaml.mustache > subgraphs/synthetix-exchanger.yaml",
+    "yaml:rates": "node ./create-yaml-file create-yaml --subgraph rates --env prod --universal-test-block null && mustache mustache/yaml_output/synthetix-rates.json mustache/yaml.mustache > subgraphs/synthetix-rates.yaml",
+    "yaml:exchanger": "node ./create-yaml-file create-yaml --subgraph exchanger --env prod --universal-test-block null && mustache mustache/yaml_output/synthetix-exchanger.json mustache/yaml.mustache > subgraphs/synthetix-exchanger.yaml",
     "codegen": "node prepare-abis && graph codegen",
     "codegen:snx": "npm run codegen -- subgraphs/synthetix.yaml",
     "codegen:rates": "npm run yaml:rates && npm run codegen -- subgraphs/synthetix-rates.yaml",
-    "codegen:rates:test": "TEST_YAML=true UNIVERSAL_START_BLOCK=null npm run yaml:rates && npm run codegen -- subgraphs/synthetix-rates.yaml",
     "codegen:depot": "npm run codegen -- subgraphs/synthetix-depot.yaml",
     "codegen:exchanges": "npm run codegen -- subgraphs/synthetix-exchanges.yaml",
     "codegen:loans": "npm run codegen -- subgraphs/synthetix-loans.yaml",
     "codegen:options": "npm run codegen -- subgraphs/synthetix-binary-options.yaml",
     "codegen:orders": "npm run codegen -- subgraphs/synthetix-limit-orders.yaml",
     "codegen:exchanger": "npm run yaml:exchanger && npm run codegen -- subgraphs/synthetix-exchanger.yaml",
-    "codegen:exchanger:test": "TEST_YAML=true UNIVERSAL_START_BLOCK=null npm run yaml:exchanger && npm run codegen -- subgraphs/synthetix-exchanger.yaml",
     "codegen:grantsdao": "npm run codegen -- subgraphs/synthetix-grantsdao.yaml",
     "codegen:liquidations": "npm run codegen -- subgraphs/synthetix-liquidations.yaml",
     "build": "graph build",
