+++ subgraphs/synthetix-rates.yaml
@@ -1624,53 +1624,3 @@ dataSources:
       eventHandlers:
         - event: AnswerUpdated(indexed int256,indexed uint256,uint256)
           handler: handleAggregatorAnswerUpdated
-
-  - kind: ethereum/contract
-    name: AggregatorOIL_3
-    network: mainnet
-    source:
-      address: '0xF0f201FDB77a5eCd065CdbD69850F3f6694B2BEf' # OIL_3
-      abi: Aggregator
-      startBlock: 11198828
-    mapping:
-      kind: ethereum/events
-      apiVersion: 0.0.4
-      language: wasm/assemblyscript
-      file: ../src/rates-mapping.ts
-      entities:
-        - AggregatorAnswer
-      abis:
-        - name: ExchangeRates
-          file: ../abis/ExchangeRates.json
-        - name: Aggregator
-          file: ../abis/Aggregator.json
-        - name: AddressResolver
-          file: ../abis/AddressResolver.json
-      eventHandlers:
-        - event: AnswerUpdated(indexed int256,indexed uint256,uint256)
-          handler: handleAggregatorAnswerUpdated
-
-  - kind: ethereum/contract
-    name: AggregatorAAVE_3
-    network: mainnet
-    source:
-      address: '0xF2d87E37EA1e54c7Aa913d2447A5f69f61C114Cf' # AAVE_3
-      abi: Aggregator
-      startBlock: 11179792
-    mapping:
-      kind: ethereum/events
-      apiVersion: 0.0.4
-      language: wasm/assemblyscript
-      file: ../src/rates-mapping.ts
-      entities:
-        - AggregatorAnswer
-      abis:
-        - name: ExchangeRates
-          file: ../abis/ExchangeRates.json
-        - name: Aggregator
-          file: ../abis/Aggregator.json
-        - name: AddressResolver
-          file: ../abis/AddressResolver.json
-      eventHandlers:
-        - event: AnswerUpdated(indexed int256,indexed uint256,uint256)
-          handler: handleAggregatorAnswerUpdated
